micro_eval/generator/requirements.txt
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.20.0
requests>=2.31.0
tqdm>=4.65.0
python-dotenv>=1.0.0
